Moved glue.qt.qtutil.load_ui to glue.utils.qt.helpers.load_ui and set up CUSTOM_QWIDGETS so any module can register new custom widgets (instead of using a hard-coded function)
