+++ glue/app/qt/plugin_manager.py
@@ -5,7 +5,7 @@ import os
 from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui
 from glue._plugin_helpers import PluginConfig
-from glue.qt.qtutil import load_ui
+from glue.utils.qt import load_ui
 
 
 __all__ = ["QtPluginManager"]
