+++ glue/core/qt/message_widget.py
@@ -4,7 +4,7 @@ from time import ctime
 
 from glue.external.qt import QtGui
 from glue import core
-from glue.qt.qtutil import load_ui
+from glue.utils.qt import load_ui
 
 
 class MessageWidget(QtGui.QWidget, core.hub.HubListener):
