+++ glue/dialogs/subset_facet/qt/subset_facet.py
@@ -6,7 +6,7 @@ from matplotlib import cm
 
 from glue.external.qt import QtGui
 from glue.core.util import colorize_subsets, facet_subsets
-from glue.qt.qtutil import load_ui
+from glue.utils.qt import load_ui
 from glue.utils.qt.widget_properties import (ButtonProperty, FloatLineProperty,
                                        ValueProperty)
 from glue.utils.array import pretty_number
