+++ glue/viewers/common/qt/mouse_mode.py
@@ -24,8 +24,9 @@ from glue.core.callback_property import CallbackProperty
 from glue.core import roi
 from glue.core.qt import roi as qt_roi
 from glue.qt import get_qapp
-from glue.qt.qtutil import get_icon, load_ui
+from glue.qt.qtutil import get_icon
 from glue.utils import nonpartial
+from glue.utils.qt import load_ui
 
 
 class MouseMode(object):
