+++ glue/viewers/histogram/qt/viewer_widget.py
@@ -9,7 +9,7 @@ from glue.core import message as msg
 from glue.viewers.histogram.client import HistogramClient
 from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.viewers.common.qt.mouse_mode import HRangeMode
-from glue.qt.qtutil import load_ui
+from glue.utils.qt import load_ui
 from glue.utils.qt.widget_properties import (connect_int_spin, ButtonProperty,
                                        FloatLineProperty, connect_float_edit,
                                        ValueProperty, connect_bool_button)
