+++ glue/viewers/scatter/qt/viewer_widget.py
@@ -9,7 +9,7 @@ from glue.viewers.scatter.client import ScatterClient
 from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.viewers.common.qt.mouse_mode import (RectangleMode, CircleMode,
                                 PolyMode, HRangeMode, VRangeMode)
-from glue.qt.qtutil import load_ui
+from glue.utils.qt import load_ui
 from glue.viewers.common.qt.data_viewer import DataViewer
 from glue.viewers.common.qt.mpl_widget import MplWidget, defer_draw
