+++ glue/viewers/table/qt/viewer_widget.py
@@ -8,7 +8,7 @@ from glue.external.qt import QtGui, QtCore
 from glue.core.subset import ElementSubsetState
 from glue.core.edit_subset_mode import EditSubsetMode
 from glue.core import message as msg
-from glue.qt.qtutil import load_ui
+from glue.utils.qt import load_ui
 from glue.viewers.common.qt.data_viewer import DataViewer
